This commit adds login functionality Created authenticate() method to handle generic authentication for username/password combination. Tests whether username or password is incorrect and replies accordingly. Max 3 attempts then back to main menu.
Also added a test method to Manager. If manager is logged in, cast a Manager* pointer to the logged in Account* and call methods from there.

Laid of generic groundwork for while loops for user/manager logins.

Added some quality of life to the text (spacing, re-ordered prompts a little bit). Just testing some stuff.